Fix/types

See merge request papers/airgap/airgap-sapling-wasm!28

+++ packages/sapling-android/app/src/main/cpp/sapling.cpp
@@ -88,7 +88,7 @@ Java_it_airgap_sapling_Sapling_extMerkleHash(
         jlong jdepth,
         jbyteArray jlhs,
         jbyteArray jrhs) {
-    auto depth = (uint64_t) jdepth;
+    auto depth = (size_t) jdepth;
 
     size_t lhs_len;
     const unsigned char *lhs = jbyteArray_to_uchar(env, jlhs, &lhs_len);
